requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Session==0.8.0
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
MarkupSafe==2.1.5
msgspec==0.18.6
ngrok==1.4.0
openai==1.40.3
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
stripe==10.10.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2